move TestApp to FullApp eliminating TestApp
