+++ src/App.jsx
@@ -2,7 +2,6 @@ import React, { Component } from 'react';
 import { HashRouter as Router, Route } from 'react-router-dom';
 import FullApp from './views/FullApp';
 import SlimApp from './views/SlimApp';
-import TestApp from './views/TestApp';
 
 class App extends Component {
 
@@ -12,7 +11,6 @@ class App extends Component {
         <div>
           <Route exact path='/' component={SlimApp} />
           <Route path='/patient' component={FullApp} />  
-		  <Route path='/test' component={TestApp} />
         </div>
       </Router>
     )
